+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - [Netkan] Add Netkan option to overwrite cached files (#2582 by HebaruSan; reviewed: politas)
 - [GUI] Show recommendations of installed modules in GUI (#2577 by HebaruSan; reviewed: politas)
 - [GUI] Remove newline from Done progress message (#2580 by HebaruSan; reviewed: politas)
+- [GUI] Set window properties on X11 (#2586 by: HebaruSan; reviewed: politas)
 
 ### Bugfixes